tests: extract core transfer tests into dedicated file
