+++ tests/test_transfer.h
@@ -1,0 +1,10 @@
+#ifndef TEST_TRANSFER_H
+#define TEST_TRANSFER_H
+
+void test_status(void);
+void test_exit(void);
+void test_switch(void);
+void test_switch_self_panic(void);
+void test_arg(void);
+
+#endif
